% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Medial_function.R
\name{medial}
\alias{medial}
\title{Medial Axis using Delaunay Triangulation}
\usage{
medial(x)
}
\arguments{
  x: A polygon or shape file; Supports convex polygons, multiple polygon, disconnected polygons, and holes in polygons
}
\description{
This function computes the medial axis for the input polygon using the Delaunay Triangulation algorithm defined by Lee and Schachter (1980). The medial axis is derived by computing the triangulations of a polygon using the package Deldir().
}
\examples{
Shape <-
  readOGR(dsn = ".\file",
          layer = "Holes")

medial(Shape)

}
\keyword{polygon}
